pyproject.toml
[project]
name = "email-mcp-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aioimaplib>=2.0.1",
    "aiosmtplib>=4.0.1",
    "apscheduler>=3.11.0",
    "fastapi>=0.115.14",
    "graphviz>=0.21",
    "langchain>=0.3.26",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.5.0",
    "mcp>=1.10.1",
    "oauthlib>=3.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests-oauthlib>=2.0.0",
    "uvicorn>=0.35.0",
]
